More refactoring

+++ Proj/AirDesk/app/src/main/java/pt/ulisboa/tecnico/cmov/airdesk/AirDeskActivity.java
@@ -11,20 +11,17 @@ import android.support.v4.app.FragmentManager;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.ActionBarActivity;
-import android.view.ContextMenu;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView.AdapterContextMenuInfo;
 import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import pt.ulisboa.tecnico.cmov.airdesk.core.workspace.Workspace;
 import pt.ulisboa.tecnico.cmov.airdesk.core.workspace.WorkspaceManager;
-import pt.ulisboa.tecnico.cmov.airdesk.fragment.CreateWorkspaceFragment;
+import pt.ulisboa.tecnico.cmov.airdesk.fragment.NavigationDrawerFragment;
 import pt.ulisboa.tecnico.cmov.airdesk.fragment.WorkspaceFragment;
 
 
@@ -178,6 +175,10 @@ public class AirDeskActivity extends ActionBarActivity implements NavigationDraw
         // automatically handle clicks on the Home/Up button, so long
         // as you specify a parent activity in AndroidManifest.xml.
 
+        if (item.getItemId() == R.id.refresh) {
+            refreshList();
+            return true;
+        }
 
         if(item.getItemId() == R.id.delete_all){
                 new AlertDialog.Builder(this)
